+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-25T09:01:09.418Z",
+    "2026-02-25T11:59:25.567Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {