+++ packages/orq-rc/RELEASES.md
@@ -7249,3 +7249,13 @@ Based on:
 - [typescript v4.5.0-rc.4] packages/orq-rc
 ### Releases
 - [NPM v4.5.0-rc.4] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.4 - packages/orq-rc
+
+## 2026-02-25 12:02:44
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.729.0 (2.841.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.5.0-rc.5] packages/orq-rc
+### Releases
+- [NPM v4.5.0-rc.5] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.5 - packages/orq-rc